src/model_tuning.py
import google.generativeai as genai
from typing import List, Dict
import os
import sys

# Add the project root to the Python path
project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
sys.path.insert(0, project_root)

from config.settings import API_KEY, MODEL_NAME, TUNED_MODEL_NAME

def setup_model():
    """Set up the Gemini model for tuning."""
    genai.configure(api_key=API_KEY)
    model = genai.GenerativeModel(MODEL_NAME)
    return model

def tune_model(model: genai.GenerativeModel, tuning_data: List[Dict[str, str]]):
    """Tune the Gemini model with the provided data."""
    # Note: The actual tuning process may differ based on the latest API
    # This is a placeholder implementation
    print("Tuning model with provided data...")
    return model  # Return the original model as a placeholder

def save_tuned_model(tuned_model, output_path: str):
    """Save the tuned model to a file."""
    # Note: Saving might not be applicable depending on how tuning works
    # This is a placeholder implementation
    print(f"Tuned model would be saved to {output_path}")